Consolidate results page banner logic
