+++ frontend/src/app/(event)/[event-code]/page-client.tsx
@@ -2,7 +2,6 @@
 
 import { useState } from "react";
 
-import { AnimatePresence, motion, Variants } from "framer-motion";
 import { PencilIcon, ShareIcon, SquarePenIcon } from "lucide-react";
 
 import CopyToastButton from "@/components/copy-toast-button";
@@ -83,64 +82,13 @@ function EventResults({ eventData }: { eventData: EventInformation }) {
   };
 
   /* BANNER */
-  const { element: banner, id: bannerId } = getResultBanner(
+  const bannerElement = getResultBanner(
     availabilities,
     participants,
     timeslots,
     eventRange.type === "weekday",
     currentUser !== null,
   );
-
-  const bannerElement = () => {
-    const variants: Variants = {
-      enter: {
-        height: "auto",
-        marginBottom: "1rem",
-        opacity: 1,
-        x: "0%",
-        transition: {
-          // Delayed extra to match with the exit animation of the banner
-          height: { duration: 0.3, delay: 0.4, ease: "easeOut" },
-          marginBottom: { duration: 0.3, delay: 0.4, ease: "easeOut" },
-          opacity: { duration: 0.4, delay: 0.7, ease: "backOut" },
-          x: { duration: 0.4, delay: 0.7, ease: "backOut" },
-        },
-      },
-      exit: {
-        height: 0,
-        marginBottom: 0,
-        opacity: 0,
-        x: "-2rem",
-        transition: {
-          opacity: { duration: 0.4, ease: "backIn" },
-          x: { duration: 0.4, ease: "backIn" },
-          height: { duration: 0.3, delay: 0.4, ease: "easeOut" },
-          marginBottom: { duration: 0.3, delay: 0.4, ease: "easeOut" },
-        },
-      },
-    };
-
-    return (
-      <AnimatePresence initial={false} mode="sync">
-        {banner && (
-          <motion.div
-            key={bannerId}
-            initial={{
-              height: 0,
-              opacity: 0,
-              x: "5%",
-              marginBottom: 0,
-            }}
-            animate="enter"
-            exit="exit"
-            variants={variants}
-          >
-            {banner}
-          </motion.div>
-        )}
-      </AnimatePresence>
-    );
-  };
 
   /* BUTTONS */
   const paintingButton = (
@@ -225,7 +173,7 @@ function EventResults({ eventData }: { eventData: EventInformation }) {
         </div>
       </div>
 
-      <div className="-mb-2 md:hidden">{bannerElement()}</div>
+      <div className="-mb-2 md:hidden">{bannerElement}</div>
 
       <div className="flex min-h-0 flex-1 flex-col md:flex-row md:gap-4">
         <ScheduleGrid
@@ -261,7 +209,7 @@ function EventResults({ eventData }: { eventData: EventInformation }) {
             "relative bottom-auto left-auto w-80 space-y-4 px-0",
           )}
         >
-          {bannerElement()}
+          {bannerElement}
           <div className="flex max-h-[calc(100vh-18rem)] flex-col gap-y-4">
             <AttendeesPanel />
             <div className="bg-panel shrink-0 rounded-3xl p-6 text-sm">
